bps: add optional child powers to inner blocks
An inner block power is amount of elements in all the leaf nodes
placed under it (through inner children if any, recursively).

Now it's possible to store child powers in inner nodes. This will
make it possible to insert/delete/select elements by offsets, but
the main goal of this is to implement logarithmic offset.

Part of #8204

NO_DOC=internal
NO_TEST=see the next commit
NO_CHANGELOG=internal