memtx: refactor the tree index lookup planning out
Before this patch the tree lookup strategy was figured out in the
`memtx_tree_index_iterator_start` function. The strategy includes
the lookup method to use, requirement to step back and the actual
lookup key and iterator type. It allows to reach a target by the
given key, iterator type and after position.

But this information will be required when fast index count will
be implemented in the memtx tree, so it needs to be refactored out
for the sake of reusability.

This patch introduces a new `memtx_tree_lookup_strategy` structure
with information required and the `memtx_tree_lookup_setup` method
to find the required strategy.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring