memtx: accelerate count in fast_offset tree
This patch creates a specialized implementation for `count` tree index
method, which takes advantage of offset-based functions of the BPS tree
with BPS_INNER_CARD config.

It also introduces the memtx_tx_index_invisible_count_matching function
which counts the amount of tuples matching with a key and iterator type
that are invisible to a transaction.

One thinig is to be noted about the memtx_tx_handle_gap_write update.

Prior to this patch, any failed lookup of a full key using EQ or REQ
iterators had been recorded using memtx_tx_track_point instead of the
memtx_tx_track_gap, so it was impossible to find a GAP_NEARBY item
with full key and EQ/REQ iterator type in the read_gaps list.

But in the fast count implementation only memtx_tx_track_gap is used
(for the sake of readability), so the memtx_tx_handle_gap_write is
modified to take it into account.

NO_DOC=TBD when fully implemented
NO_CHANGELOG=TBD when fully implemented